requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
requests
aiofiles
setuptools
wheel
